Crate nix_bindings_flake_sys

Raw bindings to Nix C API

This crate contains automatically generated bindings from the Nix C headers. The bindings are generated by bindgen and include C-style naming conventions and documentation comments that don’t always conform to Rust standards.

Normally you don’t have to use this crate directly. Instead use nix-flake.

Structs

Dl_info

Dl_serinfo

Dl_serinfo__bindgen_ty_1

Dl_serpath

StorePath

__BindgenBitfieldUnit

__BindgenUnionField

__atomic_wide_counter__bindgen_ty_1

__cancel_jmp_buf_tag

__fsid_t

__jmp_buf_tag

__kernel_fd_set

__kernel_fsid_t

__locale_data

__locale_struct

__once_flag

__pthread_cleanup_class

__pthread_cleanup_frame

__pthread_cond_s

__pthread_internal_list

__pthread_internal_slist

__pthread_mutex_s

__pthread_rwlock_arch_t

__pthread_unwind_buf_t

__sigset_t

_fpreg

_fpstate

_fpx_sw_bytes

_fpxreg

_libc_fpstate

_libc_fpxreg

_libc_xmmreg

_pthread_cleanup_buffer

_xmmreg

_xsave_hdr

_xstate

_ymmh_state

cpu_set_t

derivation

dl_find_object

flake_lock_flags

flake_reference

flake_reference_parse_flags

flake_settings

itimerspec

link_map

locked_flake

max_align_t

mcontext_t

sched_attr

sched_param

sigaction

sigcontext

sigevent

sigevent__bindgen_ty_1__bindgen_ty_1

siginfo_t

siginfo_t__bindgen_ty_1__bindgen_ty_1

siginfo_t__bindgen_ty_1__bindgen_ty_2

siginfo_t__bindgen_ty_1__bindgen_ty_3

siginfo_t__bindgen_ty_1__bindgen_ty_4

siginfo_t__bindgen_ty_1__bindgen_ty_5

siginfo_t__bindgen_ty_1__bindgen_ty_6

siginfo_t__bindgen_ty_1__bindgen_ty_7

siginfo_t__bindgen_ty_1__bindgen_ty_5__bindgen_ty_1__bindgen_ty_1

sigstack

stack_t

store_path_hash_part

@brief A store path hash

timespec

timeval

timex

tm

ucontext_t

Constants

ADJ_ESTERROR

ADJ_FREQUENCY

ADJ_MAXERROR

ADJ_MICRO

ADJ_NANO

ADJ_OFFSET

ADJ_OFFSET_SINGLESHOT

ADJ_OFFSET_SS_READ

ADJ_SETOFFSET

ADJ_STATUS

ADJ_TAI

ADJ_TICK

ADJ_TIMECONST

BUS_ADRALN

BUS_ADRERR

BUS_MCEERR_AO

BUS_MCEERR_AR

BUS_OBJERR

CLD_CONTINUED

CLD_DUMPED

CLD_EXITED

CLD_KILLED

CLD_STOPPED

CLD_TRAPPED

CLOCK_BOOTTIME

CLOCK_BOOTTIME_ALARM

CLOCK_MONOTONIC

CLOCK_MONOTONIC_COARSE

CLOCK_MONOTONIC_RAW

CLOCK_PROCESS_CPUTIME_ID

CLOCK_REALTIME

CLOCK_REALTIME_ALARM

CLOCK_REALTIME_COARSE

CLOCK_TAI

CLOCK_THREAD_CPUTIME_ID

CLONE_CHILD_CLEARTID

CLONE_CHILD_SETTID

CLONE_DETACHED

CLONE_FILES

CLONE_FS

CLONE_IO

CLONE_NEWCGROUP

CLONE_NEWIPC

CLONE_NEWNET

CLONE_NEWNS

CLONE_NEWPID

CLONE_NEWTIME

CLONE_NEWUSER

CLONE_NEWUTS

CLONE_PARENT

CLONE_PARENT_SETTID

CLONE_PIDFD

CLONE_PTRACE

CLONE_SETTLS

CLONE_SIGHAND

CLONE_SYSVSEM

CLONE_THREAD

CLONE_UNTRACED

CLONE_VFORK

CLONE_VM

CLOSE_RANGE_CLOEXEC

CLOSE_RANGE_UNSHARE

CPU_SETSIZE

CSIGNAL

DLFO_FLAG_SFRAME

DLFO_STRUCT_HAS_EH_COUNT

DLFO_STRUCT_HAS_EH_DBASE

FPE_CONDTRAP

FPE_FLTDIV

FPE_FLTINV

FPE_FLTOVF

FPE_FLTRES

FPE_FLTSUB

FPE_FLTUND

FPE_FLTUNK

FPE_INTDIV

FPE_INTOVF

FP_XSTATE_MAGIC1

FP_XSTATE_MAGIC2

F_LOCK

F_OK

F_TEST

F_TLOCK

F_ULOCK

GC_TMP_VERSION_MAJOR

GC_TMP_VERSION_MICRO

GC_TMP_VERSION_MINOR

GC_VERSION_MAJOR

GC_VERSION_MICRO

GC_VERSION_MINOR

ILL_BADIADDR

ILL_BADSTK

ILL_COPROC

ILL_ILLADR

ILL_ILLOPC

ILL_ILLOPN

ILL_ILLTRP

ILL_PRVOPC

ILL_PRVREG

INT8_MAX

INT8_MIN

INT8_WIDTH

INT16_MAX

INT16_MIN

INT16_WIDTH

INT32_MAX

INT32_MIN

INT32_WIDTH

INT64_WIDTH

INTMAX_WIDTH

INTPTR_MAX

INTPTR_MIN

INTPTR_WIDTH

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST8_WIDTH

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST16_WIDTH

INT_FAST32_MAX

INT_FAST32_MIN

INT_FAST32_WIDTH

INT_FAST64_WIDTH

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST8_WIDTH

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST16_WIDTH

INT_LEAST32_MAX

INT_LEAST32_MIN

INT_LEAST32_WIDTH

INT_LEAST64_WIDTH

LM_ID_BASE

LM_ID_NEWLM

L_INCR

L_SET

L_XTND

MINSIGSTKSZ

MOD_CLKA

MOD_CLKB

MOD_ESTERROR

MOD_FREQUENCY

MOD_MAXERROR

MOD_MICRO

MOD_NANO

MOD_OFFSET

MOD_STATUS

MOD_TAI

MOD_TIMECONST

NGREG

NSIG

POLL_ERR

POLL_HUP

POLL_IN

POLL_MSG

POLL_OUT

POLL_PRI

PTHREAD_ATTR_NO_SIGMASK_NP

PTHREAD_BARRIER_SERIAL_THREAD

PTHREAD_CANCEL_ASYNCHRONOUS

PTHREAD_CANCEL_DEFERRED

PTHREAD_CANCEL_DISABLE

PTHREAD_CANCEL_ENABLE

PTHREAD_CREATE_DETACHED

PTHREAD_CREATE_JOINABLE

PTHREAD_EXPLICIT_SCHED

PTHREAD_INHERIT_SCHED

PTHREAD_MUTEX_ADAPTIVE_NP

PTHREAD_MUTEX_DEFAULT

PTHREAD_MUTEX_ERRORCHECK

PTHREAD_MUTEX_ERRORCHECK_NP

PTHREAD_MUTEX_FAST_NP

PTHREAD_MUTEX_NORMAL

PTHREAD_MUTEX_RECURSIVE

PTHREAD_MUTEX_RECURSIVE_NP

PTHREAD_MUTEX_ROBUST

PTHREAD_MUTEX_ROBUST_NP

PTHREAD_MUTEX_STALLED

PTHREAD_MUTEX_STALLED_NP

PTHREAD_MUTEX_TIMED_NP

PTHREAD_ONCE_INIT

PTHREAD_PRIO_INHERIT

PTHREAD_PRIO_NONE

PTHREAD_PRIO_PROTECT

PTHREAD_PROCESS_PRIVATE

PTHREAD_PROCESS_SHARED

PTHREAD_RWLOCK_DEFAULT_NP

PTHREAD_RWLOCK_PREFER_READER_NP

PTHREAD_RWLOCK_PREFER_WRITER_NONRECURSIVE_NP

PTHREAD_RWLOCK_PREFER_WRITER_NP

PTHREAD_SCOPE_PROCESS

PTHREAD_SCOPE_SYSTEM

PTRDIFF_MAX

PTRDIFF_MIN

PTRDIFF_WIDTH

REG_CR2

REG_CSGSFS

REG_EFL

REG_ERR

REG_OLDMASK

REG_R8

REG_R9

REG_R10

REG_R11

REG_R12

REG_R13

REG_R14

REG_R15

REG_RAX

REG_RBP

REG_RBX

REG_RCX

REG_RDI

REG_RDX

REG_RIP

REG_RSI

REG_RSP

REG_TRAPNO

RTLD_BINDING_MASK

RTLD_DEEPBIND

RTLD_DI_CONFIGADDR

RTLD_DI_LINKMAP

RTLD_DI_LMID

RTLD_DI_MAX

RTLD_DI_ORIGIN

RTLD_DI_PHDR

RTLD_DI_PROFILENAME

RTLD_DI_PROFILEOUT

RTLD_DI_SERINFO

RTLD_DI_SERINFOSIZE

RTLD_DI_TLS_DATA

RTLD_DI_TLS_MODID

RTLD_DL_LINKMAP

RTLD_DL_SYMENT

RTLD_GLOBAL

RTLD_LAZY

RTLD_LOCAL

RTLD_NODELETE

RTLD_NOLOAD

RTLD_NOW

R_OK

SA_INTERRUPT

SA_NOCLDSTOP

SA_NOCLDWAIT

SA_NODEFER

SA_NOMASK

SA_ONESHOT

SA_ONSTACK

SA_RESETHAND

SA_RESTART

SA_SIGINFO

SA_STACK

SCHED_ATTR_SIZE_VER0

SCHED_ATTR_SIZE_VER1

SCHED_BATCH

SCHED_DEADLINE

SCHED_EXT

SCHED_FIFO

SCHED_FLAG_DL_OVERRUN

SCHED_FLAG_KEEP_ALL

SCHED_FLAG_KEEP_PARAMS

SCHED_FLAG_KEEP_POLICY

SCHED_FLAG_RECLAIM

SCHED_FLAG_RESET_ON_FORK

SCHED_FLAG_UTIL_CLAMP

SCHED_FLAG_UTIL_CLAMP_MAX

SCHED_FLAG_UTIL_CLAMP_MIN

SCHED_IDLE

SCHED_ISO

SCHED_NORMAL

SCHED_OTHER

SCHED_RESET_ON_FORK

SCHED_RR

SEEK_CUR

SEEK_DATA

SEEK_END

SEEK_HOLE

SEEK_SET

SEGV_ACCADI

SEGV_ACCERR

SEGV_ADIDERR

SEGV_ADIPERR

SEGV_BNDERR

SEGV_CPERR

SEGV_MAPERR

SEGV_MTEAERR

SEGV_MTESERR

SEGV_PKUERR

SIGABRT

SIGALRM

SIGBUS

SIGCHLD

SIGCLD

SIGCONT

SIGEV_NONE

SIGEV_SIGNAL

SIGEV_THREAD

SIGEV_THREAD_ID

SIGFPE

SIGHUP

SIGILL

SIGINT

SIGIO

SIGIOT

SIGKILL

SIGPIPE

SIGPOLL

SIGPROF

SIGPWR

SIGQUIT

SIGSEGV

SIGSTKFLT

SIGSTKSZ

SIGSTOP

SIGSYS

SIGTERM

SIGTRAP

SIGTSTP

SIGTTIN

SIGTTOU

SIGURG

SIGUSR1

SIGUSR2

SIGVTALRM

SIGWINCH

SIGXCPU

SIGXFSZ

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIG_ATOMIC_WIDTH

SIG_BLOCK

SIG_SETMASK

SIG_UNBLOCK

SIZE_MAX

SIZE_WIDTH

SI_ASYNCIO

SI_ASYNCNL

SI_DETHREAD

SI_KERNEL

SI_MESGQ

SI_QUEUE

SI_SIGIO

SI_TIMER

SI_TKILL

SI_USER

SS_DISABLE

SS_ONSTACK

STA_CLK

STA_CLOCKERR

STA_DEL

STA_FLL

STA_FREQHOLD

STA_INS

STA_MODE

STA_NANO

STA_PLL

STA_PPSERROR

STA_PPSFREQ

STA_PPSJITTER

STA_PPSSIGNAL

STA_PPSTIME

STA_PPSWANDER

STA_RONLY

STA_UNSYNC

STDERR_FILENO

STDIN_FILENO

STDOUT_FILENO

TIMER_ABSTIME

TIME_UTC

TRAP_BRANCH

TRAP_BRKPT

TRAP_HWBKPT

TRAP_TRACE

TRAP_UNK

UINT8_MAX

UINT8_WIDTH

UINT16_MAX

UINT16_WIDTH

UINT32_MAX

UINT32_WIDTH

UINT64_WIDTH

UINTMAX_WIDTH

UINTPTR_MAX

UINTPTR_WIDTH

UINT_FAST8_MAX

UINT_FAST8_WIDTH

UINT_FAST16_MAX

UINT_FAST16_WIDTH

UINT_FAST32_MAX

UINT_FAST32_WIDTH

UINT_FAST64_WIDTH

UINT_LEAST8_MAX

UINT_LEAST8_WIDTH

UINT_LEAST16_MAX

UINT_LEAST16_WIDTH

UINT_LEAST32_MAX

UINT_LEAST32_WIDTH

UINT_LEAST64_WIDTH

WCHAR_WIDTH

WINT_MAX

WINT_MIN

WINT_WIDTH

W_OK

X_OK

_ATFILE_SOURCE

_BITS_CPU_SET_H

_BITS_ENDIANNESS_H

_BITS_ENDIAN_H

_BITS_POSIX_OPT_H

_BITS_PTHREADTYPES_ARCH_H

_BITS_PTHREADTYPES_COMMON_H

_BITS_SCHED_H

_BITS_SETJMP_H

_BITS_SIGACTION_H

_BITS_SIGCONTEXT_H

_BITS_SIGEVENT_CONSTS_H

_BITS_SIGINFO_ARCH_H

_BITS_SIGINFO_CONSTS_ARCH_H

_BITS_SIGINFO_CONSTS_H

_BITS_SIGNUM_ARCH_H

_BITS_SIGNUM_GENERIC_H

_BITS_SIGSTACK_H

_BITS_SIGTHREAD_H

_BITS_SS_FLAGS_H

_BITS_STDINT_INTN_H

_BITS_STDINT_LEAST_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TIMEX_H

_BITS_TIME_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_TYPES_LOCALE_T_H

_BITS_TYPES_STRUCT_SCHED_PARAM

_BITS_TYPES___LOCALE_T_H

_BITS_WCHAR_H

_CS_GNU_LIBC_VERSION

_CS_GNU_LIBPTHREAD_VERSION

_CS_LFS64_CFLAGS

_CS_LFS64_LDFLAGS

_CS_LFS64_LIBS

_CS_LFS64_LINTFLAGS

_CS_LFS_CFLAGS

_CS_LFS_LDFLAGS

_CS_LFS_LIBS

_CS_LFS_LINTFLAGS

_CS_PATH

_CS_POSIX_V6_ILP32_OFF32_CFLAGS

_CS_POSIX_V6_ILP32_OFF32_LDFLAGS

_CS_POSIX_V6_ILP32_OFF32_LIBS

_CS_POSIX_V6_ILP32_OFF32_LINTFLAGS

_CS_POSIX_V6_ILP32_OFFBIG_CFLAGS

_CS_POSIX_V6_ILP32_OFFBIG_LDFLAGS

_CS_POSIX_V6_ILP32_OFFBIG_LIBS

_CS_POSIX_V6_ILP32_OFFBIG_LINTFLAGS

_CS_POSIX_V6_LP64_OFF64_CFLAGS

_CS_POSIX_V6_LP64_OFF64_LDFLAGS

_CS_POSIX_V6_LP64_OFF64_LIBS

_CS_POSIX_V6_LP64_OFF64_LINTFLAGS

_CS_POSIX_V6_LPBIG_OFFBIG_CFLAGS

_CS_POSIX_V6_LPBIG_OFFBIG_LDFLAGS

_CS_POSIX_V6_LPBIG_OFFBIG_LIBS

_CS_POSIX_V6_LPBIG_OFFBIG_LINTFLAGS

_CS_POSIX_V7_ILP32_OFF32_CFLAGS

_CS_POSIX_V7_ILP32_OFF32_LDFLAGS

_CS_POSIX_V7_ILP32_OFF32_LIBS

_CS_POSIX_V7_ILP32_OFF32_LINTFLAGS

_CS_POSIX_V7_ILP32_OFFBIG_CFLAGS

_CS_POSIX_V7_ILP32_OFFBIG_LDFLAGS

_CS_POSIX_V7_ILP32_OFFBIG_LIBS

_CS_POSIX_V7_ILP32_OFFBIG_LINTFLAGS

_CS_POSIX_V7_LP64_OFF64_CFLAGS

_CS_POSIX_V7_LP64_OFF64_LDFLAGS

_CS_POSIX_V7_LP64_OFF64_LIBS

_CS_POSIX_V7_LP64_OFF64_LINTFLAGS

_CS_POSIX_V7_LPBIG_OFFBIG_CFLAGS

_CS_POSIX_V7_LPBIG_OFFBIG_LDFLAGS

_CS_POSIX_V7_LPBIG_OFFBIG_LIBS

_CS_POSIX_V7_LPBIG_OFFBIG_LINTFLAGS

_CS_V5_WIDTH_RESTRICTED_ENVS

_CS_V6_ENV

_CS_V6_WIDTH_RESTRICTED_ENVS

_CS_V7_ENV

_CS_V7_WIDTH_RESTRICTED_ENVS

_CS_XBS5_ILP32_OFF32_CFLAGS

_CS_XBS5_ILP32_OFF32_LDFLAGS

_CS_XBS5_ILP32_OFF32_LIBS

_CS_XBS5_ILP32_OFF32_LINTFLAGS

_CS_XBS5_ILP32_OFFBIG_CFLAGS

_CS_XBS5_ILP32_OFFBIG_LDFLAGS

_CS_XBS5_ILP32_OFFBIG_LIBS

_CS_XBS5_ILP32_OFFBIG_LINTFLAGS

_CS_XBS5_LP64_OFF64_CFLAGS

_CS_XBS5_LP64_OFF64_LDFLAGS

_CS_XBS5_LP64_OFF64_LIBS

_CS_XBS5_LP64_OFF64_LINTFLAGS

_CS_XBS5_LPBIG_OFFBIG_CFLAGS

_CS_XBS5_LPBIG_OFFBIG_LDFLAGS

_CS_XBS5_LPBIG_OFFBIG_LIBS

_CS_XBS5_LPBIG_OFFBIG_LINTFLAGS

_DEFAULT_SOURCE

_DLFCN_H

_DYNAMIC_STACK_SIZE_SOURCE

_FEATURES_H

_GETOPT_CORE_H

_GETOPT_POSIX_H

_ISOC2Y_SOURCE

_ISOC11_SOURCE

_ISOC23_SOURCE

_ISOC95_SOURCE

_ISOC99_SOURCE

_LARGEFILE64_SOURCE

_LARGEFILE_SOURCE

_LFS64_ASYNCHRONOUS_IO

_LFS64_LARGEFILE

_LFS64_STDIO

_LFS_ASYNCHRONOUS_IO

_LFS_LARGEFILE

_NSIG

_PC_2_SYMLINKS

_PC_ALLOC_SIZE_MIN

_PC_ASYNC_IO

_PC_CHOWN_RESTRICTED

_PC_FILESIZEBITS

_PC_LINK_MAX

_PC_MAX_CANON

_PC_MAX_INPUT

_PC_NAME_MAX

_PC_NO_TRUNC

_PC_PATH_MAX

_PC_PIPE_BUF

_PC_PRIO_IO

_PC_REC_INCR_XFER_SIZE

_PC_REC_MAX_XFER_SIZE

_PC_REC_MIN_XFER_SIZE

_PC_REC_XFER_ALIGN

_PC_SOCK_MAXBUF

_PC_SYMLINK_MAX

_PC_SYNC_IO

_PC_VDISABLE

_POSIX2_CHAR_TERM

_POSIX2_C_BIND

_POSIX2_C_DEV

_POSIX2_C_VERSION

_POSIX2_LOCALEDEF

_POSIX2_SW_DEV

_POSIX2_VERSION

_POSIX_ADVISORY_INFO

_POSIX_ASYNCHRONOUS_IO

_POSIX_ASYNC_IO

_POSIX_BARRIERS

_POSIX_CHOWN_RESTRICTED

_POSIX_CLOCK_SELECTION

_POSIX_CPUTIME

_POSIX_C_SOURCE

_POSIX_FSYNC

_POSIX_IPV6

_POSIX_JOB_CONTROL

_POSIX_MAPPED_FILES

_POSIX_MEMLOCK

_POSIX_MEMLOCK_RANGE

_POSIX_MEMORY_PROTECTION

_POSIX_MESSAGE_PASSING

_POSIX_MONOTONIC_CLOCK

_POSIX_NO_TRUNC

_POSIX_PRIORITIZED_IO

_POSIX_PRIORITY_SCHEDULING

_POSIX_RAW_SOCKETS

_POSIX_READER_WRITER_LOCKS

_POSIX_REALTIME_SIGNALS

_POSIX_REENTRANT_FUNCTIONS

_POSIX_REGEXP

_POSIX_SAVED_IDS

_POSIX_SEMAPHORES

_POSIX_SHARED_MEMORY_OBJECTS

_POSIX_SHELL

_POSIX_SOURCE

_POSIX_SPAWN

_POSIX_SPIN_LOCKS

_POSIX_SPORADIC_SERVER

_POSIX_SYNCHRONIZED_IO

_POSIX_THREADS

_POSIX_THREAD_ATTR_STACKADDR

_POSIX_THREAD_ATTR_STACKSIZE

_POSIX_THREAD_CPUTIME

_POSIX_THREAD_PRIORITY_SCHEDULING

_POSIX_THREAD_PRIO_INHERIT

_POSIX_THREAD_PRIO_PROTECT

_POSIX_THREAD_PROCESS_SHARED

_POSIX_THREAD_ROBUST_PRIO_INHERIT

_POSIX_THREAD_ROBUST_PRIO_PROTECT

_POSIX_THREAD_SAFE_FUNCTIONS

_POSIX_THREAD_SPORADIC_SERVER

_POSIX_TIMEOUTS

_POSIX_TIMERS

_POSIX_TRACE

_POSIX_TRACE_EVENT_FILTER

_POSIX_TRACE_INHERIT

_POSIX_TRACE_LOG

_POSIX_TYPED_MEMORY_OBJECTS

_POSIX_V6_LP64_OFF64

_POSIX_V6_LPBIG_OFFBIG

_POSIX_V7_LP64_OFF64

_POSIX_V7_LPBIG_OFFBIG

_POSIX_VDISABLE

_POSIX_VERSION

_PTHREAD_H

_REENTRANT

_SCHED_H

_SC_2_CHAR_TERM

_SC_2_C_BIND

_SC_2_C_DEV

_SC_2_C_VERSION

_SC_2_FORT_DEV

_SC_2_FORT_RUN

_SC_2_LOCALEDEF

_SC_2_PBS

_SC_2_PBS_ACCOUNTING

_SC_2_PBS_CHECKPOINT

_SC_2_PBS_LOCATE

_SC_2_PBS_MESSAGE

_SC_2_PBS_TRACK

_SC_2_SW_DEV

_SC_2_UPE

_SC_2_VERSION

_SC_ADVISORY_INFO

_SC_AIO_LISTIO_MAX

_SC_AIO_MAX

_SC_AIO_PRIO_DELTA_MAX

_SC_ARG_MAX

_SC_ASYNCHRONOUS_IO

_SC_ATEXIT_MAX

_SC_AVPHYS_PAGES

_SC_BARRIERS

_SC_BASE

_SC_BC_BASE_MAX

_SC_BC_DIM_MAX

_SC_BC_SCALE_MAX

_SC_BC_STRING_MAX

_SC_CHARCLASS_NAME_MAX

_SC_CHAR_BIT

_SC_CHAR_MAX

_SC_CHAR_MIN

_SC_CHILD_MAX

_SC_CLK_TCK

_SC_CLOCK_SELECTION

_SC_COLL_WEIGHTS_MAX

_SC_CPUTIME

_SC_C_LANG_SUPPORT

_SC_C_LANG_SUPPORT_R

_SC_DELAYTIMER_MAX

_SC_DEVICE_IO

_SC_DEVICE_SPECIFIC

_SC_DEVICE_SPECIFIC_R

_SC_EQUIV_CLASS_MAX

_SC_EXPR_NEST_MAX

_SC_FD_MGMT

_SC_FIFO

_SC_FILE_ATTRIBUTES

_SC_FILE_LOCKING

_SC_FILE_SYSTEM

_SC_FSYNC

_SC_GETGR_R_SIZE_MAX

_SC_GETPW_R_SIZE_MAX

_SC_HOST_NAME_MAX

_SC_INT_MAX

_SC_INT_MIN

_SC_IOV_MAX

_SC_IPV6

_SC_JOB_CONTROL

_SC_LEVEL1_DCACHE_ASSOC

_SC_LEVEL1_DCACHE_LINESIZE

_SC_LEVEL1_DCACHE_SIZE

_SC_LEVEL1_ICACHE_ASSOC

_SC_LEVEL1_ICACHE_LINESIZE

_SC_LEVEL1_ICACHE_SIZE

_SC_LEVEL2_CACHE_ASSOC

_SC_LEVEL2_CACHE_LINESIZE

_SC_LEVEL2_CACHE_SIZE

_SC_LEVEL3_CACHE_ASSOC

_SC_LEVEL3_CACHE_LINESIZE

_SC_LEVEL3_CACHE_SIZE

_SC_LEVEL4_CACHE_ASSOC

_SC_LEVEL4_CACHE_LINESIZE

_SC_LEVEL4_CACHE_SIZE

_SC_LINE_MAX

_SC_LOGIN_NAME_MAX

_SC_LONG_BIT

_SC_MAPPED_FILES

_SC_MB_LEN_MAX

_SC_MEMLOCK

_SC_MEMLOCK_RANGE

_SC_MEMORY_PROTECTION

_SC_MESSAGE_PASSING

_SC_MINSIGSTKSZ

_SC_MONOTONIC_CLOCK

_SC_MQ_OPEN_MAX

_SC_MQ_PRIO_MAX

_SC_MULTI_PROCESS

_SC_NETWORKING

_SC_NGROUPS_MAX

_SC_NL_ARGMAX

_SC_NL_LANGMAX

_SC_NL_MSGMAX

_SC_NL_NMAX

_SC_NL_SETMAX

_SC_NL_TEXTMAX

_SC_NPROCESSORS_CONF

_SC_NPROCESSORS_ONLN

_SC_NZERO

_SC_OPEN_MAX

_SC_PAGESIZE

_SC_PASS_MAX

_SC_PHYS_PAGES

_SC_PII

_SC_PII_INTERNET

_SC_PII_INTERNET_DGRAM

_SC_PII_INTERNET_STREAM

_SC_PII_OSI

_SC_PII_OSI_CLTS

_SC_PII_OSI_COTS

_SC_PII_OSI_M

_SC_PII_SOCKET

_SC_PII_XTI

_SC_PIPE

_SC_POLL

_SC_PRIORITIZED_IO

_SC_PRIORITY_SCHEDULING

_SC_RAW_SOCKETS

_SC_READER_WRITER_LOCKS

_SC_REALTIME_SIGNALS

_SC_REGEXP

_SC_REGEX_VERSION

_SC_RE_DUP_MAX

_SC_RTSIG_MAX

_SC_SAVED_IDS

_SC_SCHAR_MAX

_SC_SCHAR_MIN

_SC_SELECT

_SC_SEMAPHORES

_SC_SEM_NSEMS_MAX

_SC_SEM_VALUE_MAX

_SC_SHARED_MEMORY_OBJECTS

_SC_SHELL

_SC_SHRT_MAX

_SC_SHRT_MIN

_SC_SIGNALS

_SC_SIGQUEUE_MAX

_SC_SIGSTKSZ

_SC_SINGLE_PROCESS

_SC_SPAWN

_SC_SPIN_LOCKS

_SC_SPORADIC_SERVER

_SC_SSIZE_MAX

_SC_SS_REPL_MAX

_SC_STREAMS

_SC_STREAM_MAX

_SC_SYMLOOP_MAX

_SC_SYNCHRONIZED_IO

_SC_SYSTEM_DATABASE

_SC_SYSTEM_DATABASE_R

_SC_THREADS

_SC_THREAD_ATTR_STACKADDR

_SC_THREAD_ATTR_STACKSIZE

_SC_THREAD_CPUTIME

_SC_THREAD_DESTRUCTOR_ITERATIONS

_SC_THREAD_KEYS_MAX

_SC_THREAD_PRIORITY_SCHEDULING

_SC_THREAD_PRIO_INHERIT

_SC_THREAD_PRIO_PROTECT

_SC_THREAD_PROCESS_SHARED

_SC_THREAD_ROBUST_PRIO_INHERIT

_SC_THREAD_ROBUST_PRIO_PROTECT

_SC_THREAD_SAFE_FUNCTIONS

_SC_THREAD_SPORADIC_SERVER

_SC_THREAD_STACK_MIN

_SC_THREAD_THREADS_MAX

_SC_TIMEOUTS

_SC_TIMERS

_SC_TIMER_MAX

_SC_TRACE

_SC_TRACE_EVENT_FILTER

_SC_TRACE_EVENT_NAME_MAX

_SC_TRACE_INHERIT

_SC_TRACE_LOG

_SC_TRACE_NAME_MAX

_SC_TRACE_SYS_MAX

_SC_TRACE_USER_EVENT_MAX

_SC_TTY_NAME_MAX

_SC_TYPED_MEMORY_OBJECTS

_SC_TZNAME_MAX

_SC_T_IOV_MAX

_SC_UCHAR_MAX

_SC_UINT_MAX

_SC_UIO_MAXIOV

_SC_ULONG_MAX

_SC_USER_GROUPS

_SC_USER_GROUPS_R

_SC_USHRT_MAX

_SC_V6_ILP32_OFF32

_SC_V6_ILP32_OFFBIG

_SC_V6_LP64_OFF64

_SC_V6_LPBIG_OFFBIG

_SC_V7_ILP32_OFF32

_SC_V7_ILP32_OFFBIG

_SC_V7_LP64_OFF64

_SC_V7_LPBIG_OFFBIG

_SC_VERSION

_SC_WORD_BIT

_SC_XBS5_ILP32_OFF32

_SC_XBS5_ILP32_OFFBIG

_SC_XBS5_LP64_OFF64

_SC_XBS5_LPBIG_OFFBIG

_SC_XOPEN_CRYPT

_SC_XOPEN_ENH_I18N

_SC_XOPEN_LEGACY

_SC_XOPEN_REALTIME

_SC_XOPEN_REALTIME_THREADS

_SC_XOPEN_SHM

_SC_XOPEN_STREAMS

_SC_XOPEN_UNIX

_SC_XOPEN_VERSION

_SC_XOPEN_XCU_VERSION

_SC_XOPEN_XPG2

_SC_XOPEN_XPG3

_SC_XOPEN_XPG4

_STDC_PREDEF_H

_STDINT_H

_STRUCT_TIMESPEC

_SYS_CDEFS_H

_SYS_UCONTEXT_H

_THREAD_MUTEX_INTERNAL_H

_THREAD_SHARED_TYPES_H

_TIME_H

_UNISTD_H

_XBS5_LP64_OFF64

_XBS5_LPBIG_OFFBIG

_XOPEN_ENH_I18N

_XOPEN_LEGACY

_XOPEN_REALTIME

_XOPEN_REALTIME_THREADS

_XOPEN_SHM

_XOPEN_SOURCE

_XOPEN_SOURCE_EXTENDED

_XOPEN_UNIX

_XOPEN_VERSION

_XOPEN_XCU_VERSION

_XOPEN_XPG2

_XOPEN_XPG3

_XOPEN_XPG4

__BIG_ENDIAN

__BITS_PER_LONG

__BITS_PER_LONG_LONG

__BYTE_ORDER

__CPU_SETSIZE

__FD_SETSIZE

__FLOAT_WORD_ORDER

__GLIBC_MINOR__

__GLIBC_USE_C23_STRTOL

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C23

__GLIBC_USE_IEC_60559_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C23

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2Y

__GLIBC_USE_ISOC23

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__ILP32_OFF32_CFLAGS

__ILP32_OFF32_LDFLAGS

__ILP32_OFFBIG_CFLAGS

__ILP32_OFFBIG_LDFLAGS

__INO_T_MATCHES_INO64_T

__KERNEL_OLD_TIMEVAL_MATCHES_TIMEVAL64

__LDOUBLE_REDIRECTS_TO_FLOAT128_ABI

__LITTLE_ENDIAN

__LP64_OFF64_CFLAGS

__LP64_OFF64_LDFLAGS

__NGREG

__OFF_T_MATCHES_OFF64_T

__PDP_ENDIAN

__POSIX2_THIS_VERSION

__PTHREAD_MUTEX_HAVE_PREV

__RLIM_T_MATCHES_RLIM64_T

__SC_THREAD_STACK_MIN_VALUE

__SIGEV_MAX_SIZE

__SIGRTMAX

__SIGRTMIN

__SIZEOF_PTHREAD_ATTR_T

__SIZEOF_PTHREAD_BARRIERATTR_T

__SIZEOF_PTHREAD_BARRIER_T

__SIZEOF_PTHREAD_CONDATTR_T

__SIZEOF_PTHREAD_COND_T

__SIZEOF_PTHREAD_MUTEXATTR_T

__SIZEOF_PTHREAD_MUTEX_T

__SIZEOF_PTHREAD_RWLOCKATTR_T

__SIZEOF_PTHREAD_RWLOCK_T

__SI_ASYNCIO_AFTER_SIGIO

__SI_ERRNO_THEN_CODE

__SI_HAVE_SIGSYS

__SI_MAX_SIZE

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_IEC_60559_BFP__

__STDC_IEC_60559_COMPLEX__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_DYNAMIC_STACK_SIZE

__USE_FORTIFY_LEVEL

__USE_GNU

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_ISOCXX11

__USE_LARGEFILE

__USE_LARGEFILE64

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_TIME_BITS64

__USE_UNIX98

__USE_XOPEN

__USE_XOPEN2K

__USE_XOPEN2K8

__USE_XOPEN2K8XSI

__USE_XOPEN2KXSI

__USE_XOPEN_EXTENDED

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__bool_true_false_are_defined

__clock_t_defined

__clockid_t_defined

__glibc_c99_flexarr_available

__have_pthread_attr_t

__itimerspec_defined

__jmp_buf_tag_defined

__sig_atomic_t_defined

__sigevent_t_defined

__siginfo_t_defined

__sigset_t_defined

__sigstack_defined

__stack_t_defined

__struct_tm_defined

__time_t_defined

__timer_t_defined

__timeval_defined

Statics

__daylight⚠

__environ⚠

__timezone⚠

__tzname⚠

daylight⚠

environ⚠

getdate_err⚠

optarg⚠

opterr⚠

optind⚠

optopt⚠

timezone⚠

tzname⚠

Functions

GC_dlopen⚠

GC_pthread_cancel⚠

GC_pthread_create⚠

GC_pthread_detach⚠

GC_pthread_exit⚠

GC_pthread_join⚠

GC_pthread_sigmask⚠

_Fork⚠

__getpgid⚠

__libc_current_sigrtmax⚠

__libc_current_sigrtmin⚠

__sched_cpualloc⚠

__sched_cpucount⚠

__sched_cpufree⚠

__sigsetjmp⚠

__sysconf⚠

__sysv_signal⚠

_dl_find_object⚠

_dl_mcount_wrapper_check⚠

_exit⚠

access⚠

acct⚠

alarm⚠

asctime⚠

asctime_r⚠

brk⚠

chdir⚠

chown⚠

chroot⚠

clock⚠

clock_adjtime⚠

clock_getcpuclockid⚠

clock_getres⚠

clock_gettime⚠

clock_nanosleep⚠

clock_settime⚠

clone⚠

close⚠

close_range⚠

closefrom⚠

confstr⚠

copy_file_range⚠

crypt⚠

ctime⚠

ctime_r⚠

daemon⚠

derivation_clone⚠

@brief Copy a nix_derivation

derivation_free⚠

@brief Deallocate a nix_derivation

derivation_to_json⚠

@brief Gets the derivation as a JSON string

difftime⚠

dladdr⚠

dladdr1⚠

dlclose⚠

dlerror⚠

dlinfo⚠

dlmopen⚠

dlopen⚠

dlsym⚠

dlvsym⚠

dup⚠

dup2⚠

dup3⚠

dysize⚠

eaccess⚠

endusershell⚠

euidaccess⚠

execl⚠

execle⚠

execlp⚠

execv⚠

execve⚠

execveat⚠

execvp⚠

execvpe⚠

faccessat⚠

fchdir⚠

fchown⚠

fchownat⚠

fdatasync⚠

fexecve⚠

flake_lock⚠

@brief Lock a flake, if not already locked. @param[out] context Optional, stores error information @param[in] settings The flake (and fetch) settings to use @param[in] flags The locking flags to use @param[in] flake The flake to lock

flake_lock_flags_add_input_override⚠

@brief Add input overrides to the lock flags @param[out] context Optional, stores error information @param[in] flags The flags to modify @param[in] inputPath The input path to override (must not be empty) @param[in] flakeRef The flake reference to use as the override @return NIX_ERR_NIX_ERROR if inputPath is empty

flake_lock_flags_free⚠

@brief Deallocate and release the resources associated with a nix_flake_lock_flags. Does not fail. @param[in] settings the nix_flake_lock_flags * to free

flake_lock_flags_new⚠

@brief A new nix_flake_lock_flags with defaults @param[in] settings Flake settings that may affect the defaults

flake_lock_flags_set_mode_check⚠

@brief Put the lock flags in a mode that checks whether the lock is up to date. @param[out] context Optional, stores error information @param[in] flags The flags to modify @return NIX_OK on success, NIX_ERR on failure

flake_lock_flags_set_mode_virtual⚠

@brief Put the lock flags in a mode that updates the lock file in memory, if needed. @param[out] context Optional, stores error information @param[in] flags The flags to modify @param[in] update Whether to allow updates

flake_lock_flags_set_mode_write_as_needed⚠

@brief Put the lock flags in a mode that updates the lock file on disk, if needed. @param[out] context Optional, stores error information @param[in] flags The flags to modify @param[in] update Whether to allow updates

flake_reference_and_fragment_from_string⚠

@brief Parse a URL-like string into a nix_flake_reference.

flake_reference_free⚠

@brief Deallocate and release the resources associated with a nix_flake_reference.

flake_reference_parse_flags_free⚠

@brief Deallocate and release the resources associated with a nix_flake_reference_parse_flags. Does not fail. @param[in] flags the nix_flake_reference_parse_flags * to free

flake_reference_parse_flags_new⚠

@brief A new nix_flake_reference_parse_flags with defaults

flake_reference_parse_flags_set_base_directory⚠

@brief Provide a base directory for parsing relative flake references @param[out] context Optional, stores error information @param[in] flags The flags to modify @param[in] baseDirectory The base directory to add @param[in] baseDirectoryLen The length of baseDirectory @return NIX_OK on success, NIX_ERR on failure

flake_settings_add_to_eval_state_builder⚠

@brief Initialize a nix_flake_settings to contain builtins.getFlake and potentially more.

flake_settings_free⚠

@brief Release the resources associated with a nix_flake_settings.

flake_settings_new⚠

Create a nix_flake_settings initialized with default values. @param[out] context Optional, stores error information @return A new nix_flake_settings or NULL on failure. @see nix_flake_settings_free

fork⚠

fpathconf⚠

fsync⚠

ftruncate⚠

ftruncate64⚠

get_current_dir_name⚠

getcpu⚠

getcwd⚠

getdate⚠

getdate_r⚠

getdomainname⚠

getdtablesize⚠

getegid⚠

getentropy⚠

geteuid⚠

getgid⚠

getgroups⚠

gethostid⚠

gethostname⚠

getlogin⚠

getlogin_r⚠

getopt⚠

getpagesize⚠

getpass⚠

getpgid⚠

getpgrp⚠

getpid⚠

getppid⚠

getresgid⚠

getresuid⚠

getsid⚠

gettid⚠

getuid⚠

getusershell⚠

getwd⚠

gmtime⚠

gmtime_r⚠

group_member⚠

gsignal⚠

isatty⚠

kill⚠

killpg⚠

lchown⚠

link⚠

linkat⚠

localtime⚠

localtime_r⚠

locked_flake_free⚠

@brief Deallocate and release the resources associated with a nix_locked_flake. Does not fail. @param[in] locked_flake the nix_locked_flake * to free

locked_flake_get_output_attrs⚠

@brief Get the output attributes of a flake. @param[out] context Optional, stores error information @param[in] settings The settings to use @param[in] locked_flake the flake to get the output attributes from @return A new nix_value or NULL on failure. Release the nix_value with nix_value_decref.

lockf⚠

lockf64⚠

lseek⚠

lseek64⚠

mktime⚠

nanosleep⚠

nice⚠

pathconf⚠

pause⚠

pipe⚠

pipe2⚠

pread⚠

pread64⚠

profil⚠

psiginfo⚠

psignal⚠

pthread_atfork⚠

pthread_attr_destroy⚠

pthread_attr_getaffinity_np⚠

pthread_attr_getdetachstate⚠

pthread_attr_getguardsize⚠

pthread_attr_getinheritsched⚠

pthread_attr_getschedparam⚠

pthread_attr_getschedpolicy⚠

pthread_attr_getscope⚠

pthread_attr_getsigmask_np⚠

pthread_attr_getstack⚠

pthread_attr_getstackaddr⚠

pthread_attr_getstacksize⚠

pthread_attr_init⚠

pthread_attr_setaffinity_np⚠

pthread_attr_setdetachstate⚠

pthread_attr_setguardsize⚠

pthread_attr_setinheritsched⚠

pthread_attr_setschedparam⚠

pthread_attr_setschedpolicy⚠

pthread_attr_setscope⚠

pthread_attr_setsigmask_np⚠

pthread_attr_setstack⚠

pthread_attr_setstackaddr⚠

pthread_attr_setstacksize⚠

pthread_barrier_destroy⚠

pthread_barrier_init⚠

pthread_barrier_wait⚠

pthread_barrierattr_destroy⚠

pthread_barrierattr_getpshared⚠

pthread_barrierattr_init⚠

pthread_barrierattr_setpshared⚠

pthread_cancel⚠

pthread_clockjoin_np⚠

pthread_cond_broadcast⚠

pthread_cond_clockwait⚠

pthread_cond_destroy⚠

pthread_cond_init⚠

pthread_cond_signal⚠

pthread_cond_timedwait⚠

pthread_cond_wait⚠

pthread_condattr_destroy⚠

pthread_condattr_getclock⚠

pthread_condattr_getpshared⚠

pthread_condattr_init⚠

pthread_condattr_setclock⚠

pthread_condattr_setpshared⚠

pthread_create⚠

pthread_detach⚠

pthread_equal⚠

pthread_exit⚠

pthread_getaffinity_np⚠

pthread_getattr_default_np⚠

pthread_getattr_np⚠

pthread_getconcurrency⚠

pthread_getcpuclockid⚠

pthread_getname_np⚠

pthread_getschedparam⚠

pthread_getspecific⚠

pthread_gettid_np⚠

pthread_join⚠

pthread_key_create⚠

pthread_key_delete⚠

pthread_kill⚠

pthread_mutex_clocklock⚠

pthread_mutex_consistent⚠

pthread_mutex_destroy⚠

pthread_mutex_getprioceiling⚠

pthread_mutex_init⚠

pthread_mutex_lock⚠

pthread_mutex_setprioceiling⚠

pthread_mutex_timedlock⚠

pthread_mutex_trylock⚠

pthread_mutex_unlock⚠

pthread_mutexattr_destroy⚠

pthread_mutexattr_getprioceiling⚠

pthread_mutexattr_getprotocol⚠

pthread_mutexattr_getpshared⚠

pthread_mutexattr_getrobust⚠

pthread_mutexattr_gettype⚠

pthread_mutexattr_init⚠

pthread_mutexattr_setprioceiling⚠

pthread_mutexattr_setprotocol⚠

pthread_mutexattr_setpshared⚠

pthread_mutexattr_setrobust⚠

pthread_mutexattr_settype⚠

pthread_once⚠

pthread_rwlock_clockrdlock⚠

pthread_rwlock_clockwrlock⚠

pthread_rwlock_destroy⚠

pthread_rwlock_init⚠

pthread_rwlock_rdlock⚠

pthread_rwlock_timedrdlock⚠

pthread_rwlock_timedwrlock⚠

pthread_rwlock_tryrdlock⚠

pthread_rwlock_trywrlock⚠

pthread_rwlock_unlock⚠

pthread_rwlock_wrlock⚠

pthread_rwlockattr_destroy⚠

pthread_rwlockattr_getkind_np⚠

pthread_rwlockattr_getpshared⚠

pthread_rwlockattr_init⚠

pthread_rwlockattr_setkind_np⚠

pthread_rwlockattr_setpshared⚠

pthread_self⚠

pthread_setaffinity_np⚠

pthread_setattr_default_np⚠

pthread_setcancelstate⚠

pthread_setcanceltype⚠

pthread_setconcurrency⚠

pthread_setname_np⚠

pthread_setschedparam⚠

pthread_setschedprio⚠

pthread_setspecific⚠

pthread_sigmask⚠

pthread_sigqueue⚠

pthread_spin_destroy⚠

pthread_spin_init⚠

pthread_spin_lock⚠

pthread_spin_trylock⚠

pthread_spin_unlock⚠

pthread_testcancel⚠

pthread_timedjoin_np⚠

pthread_tryjoin_np⚠

pthread_yield⚠

pwrite⚠

pwrite64⚠

raise⚠

read⚠

readlink⚠

readlinkat⚠

revoke⚠

rmdir⚠

sbrk⚠

sched_get_priority_max⚠

sched_get_priority_min⚠

sched_getaffinity⚠

sched_getattr⚠

sched_getcpu⚠

sched_getparam⚠

sched_getscheduler⚠

sched_rr_get_interval⚠

sched_setaffinity⚠

sched_setattr⚠

sched_setparam⚠

sched_setscheduler⚠

sched_yield⚠

setdomainname⚠

setegid⚠

seteuid⚠

setgid⚠

sethostid⚠

sethostname⚠

setlogin⚠

setns⚠

setpgid⚠

setpgrp⚠

setregid⚠

setresgid⚠

setresuid⚠

setreuid⚠

setsid⚠

setuid⚠

setusershell⚠

sigaction⚠

sigaddset⚠

sigaltstack⚠

sigandset⚠

sigblock⚠

sigdelset⚠

sigemptyset⚠

sigfillset⚠

siggetmask⚠

sighold⚠

sigignore⚠

siginterrupt⚠

sigisemptyset⚠

sigismember⚠

signal⚠

sigorset⚠

sigpause⚠

sigpending⚠

sigprocmask⚠

sigqueue⚠

sigrelse⚠

sigreturn⚠

sigset⚠

sigsetmask⚠

sigstack⚠

sigsuspend⚠

sigtimedwait⚠

sigwait⚠

sigwaitinfo⚠

sleep⚠

ssignal⚠

store_create_from_parts⚠

@brief Create a StorePath from its constituent parts (hash and name)

store_path_clone⚠

@brief Copy a StorePath

store_path_free⚠

@brief Deallocate a StorePath

store_path_hash⚠

@brief Get the path hash (e.g. “” in /nix/store/-)

store_path_name⚠

@brief Get the path name (e.g. “” in /nix/store/-)

strftime⚠

strftime_l⚠

strptime⚠

strptime_l⚠

swab⚠

symlink⚠

symlinkat⚠

sync⚠

syncfs⚠

syscall⚠

sysconf⚠

sysv_signal⚠

tcgetpgrp⚠

tcsetpgrp⚠

tgkill⚠

time⚠

timegm⚠

timelocal⚠

timer_create⚠

timer_delete⚠

timer_getoverrun⚠

timer_gettime⚠

timer_settime⚠

timespec_get⚠

timespec_getres⚠

truncate⚠

truncate64⚠

ttyname⚠

ttyname_r⚠

ttyslot⚠

tzset⚠

ualarm⚠

unlink⚠

unlinkat⚠

unshare⚠

usleep⚠

vfork⚠

vhangup⚠

write⚠

Type Aliases

Lmid_t

__be16

__be32

__be64

__blkcnt64_t

__blkcnt_t

__blksize_t

__caddr_t

__clock_t

__clockid_t

__cpu_mask

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__jmp_buf

__kernel_caddr_t

__kernel_clock_t

__kernel_clockid_t

__kernel_daddr_t

__kernel_gid16_t

__kernel_gid32_t

__kernel_gid_t

__kernel_ino_t

__kernel_ipc_pid_t

__kernel_key_t

__kernel_loff_t

__kernel_long_t

__kernel_mode_t

__kernel_mqd_t

__kernel_off_t

__kernel_old_dev_t

__kernel_old_gid_t

__kernel_old_time_t

__kernel_old_uid_t

__kernel_pid_t

__kernel_ptrdiff_t

__kernel_sighandler_t

__kernel_size_t

__kernel_ssize_t

__kernel_suseconds_t

__kernel_time64_t

__kernel_time_t

__kernel_timer_t

__kernel_uid16_t

__kernel_uid32_t

__kernel_uid_t

__kernel_ulong_t

__key_t

__le16

__le32

__le64

__locale_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__poll_t

__pthread_list_t

__pthread_slist_t

__quad_t

__rlim64_t

__rlim_t

__s8

__s16

__s32

__s64

__s128

__sig_atomic_t

__sighandler_t

__sigval_t

__socklen_t

__ssize_t

__sum16

__suseconds64_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__thrd_t

__time_t

__timer_t

__tss_t

__u8

__u16

__u32

__u64

__u128

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

__wsum

_bindgen_ty_1

_bindgen_ty_2

_bindgen_ty_3

_bindgen_ty_4

_bindgen_ty_5

_bindgen_ty_6

_bindgen_ty_7

_bindgen_ty_8

_bindgen_ty_9

_bindgen_ty_10

_bindgen_ty_11

_bindgen_ty_12

_bindgen_ty_13

_bindgen_ty_14

_bindgen_ty_15

_bindgen_ty_16

_bindgen_ty_17

_bindgen_ty_18

_bindgen_ty_19

_bindgen_ty_20

_bindgen_ty_21

_bindgen_ty_22

_bindgen_ty_23

_bindgen_ty_24

_bindgen_ty_25

_bindgen_ty_26

clock_t

clockid_t

fpregset_t

gid_t

greg_t

gregset_t

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

locale_t

off64_t

off_t

pid_t

pthread_key_t

pthread_once_t

pthread_spinlock_t

pthread_t

sig_atomic_t

sig_t

sigevent_t

sighandler_t

sigset_t

sigval_t

socklen_t

time_t

timer_t

uid_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

useconds_t

Unions

__atomic_wide_counter

pthread_attr_t

pthread_barrier_t

pthread_barrierattr_t

pthread_cond_t

pthread_condattr_t

pthread_mutex_t

pthread_mutexattr_t

pthread_rwlock_t

pthread_rwlockattr_t

sigaction__bindgen_ty_1

sigcontext__bindgen_ty_1

sigevent__bindgen_ty_1

siginfo_t__bindgen_ty_1

siginfo_t__bindgen_ty_1__bindgen_ty_5__bindgen_ty_1

sigval