nix_bindings_flake_sys

Function store_path_hash

pub unsafe extern "C" fn store_path_hash(
    context: *mut c_context,
    store_path: *const StorePath,
    hash_part_out: *mut store_path_hash_part,
) -> err

@brief Get the path hash (e.g. “” in /nix/store/-)

The hash is returned as raw bytes, decoded from “nix32” encoding.

@param[out] context Optional, stores error information @param[in] store_path the path to get the hash from @param[out] hash_part_out the decoded hash as 20 raw bytes @return NIX_OK on success, error code on failure